% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{composition}
\alias{composition}
\title{Retrieve one or more product compositions}
\usage{
composition(pNbrs)
}
\arguments{
\item{pNbrs}{A numeric vector of one or more P-Numbers}
}
\description{
Retrieve one or more product compositions
}
