% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{dummy_data.tree}
\alias{dummy_data.tree}
\title{Dummy data.tree node included to allow data.tree in Imports}
\format{
An object of class \code{Node} (inherits from \code{R6}) of length 41.
}
\usage{
dummy_data.tree
}
\description{
Dummy data.tree node included to allow data.tree in Imports
}
\keyword{internal}
