% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_approved.R
\name{last_approved}
\alias{last_approved}
\title{Find the year of the last approved use of an active substance}
\usage{
last_approved(pattern, lang = c("de", "fr", "it"))
}
\arguments{
\item{pattern}{Search pattern for the substance name, passed to \link{grepl}.}

\item{lang}{Should the German names be searched in (default), or the French
or the Italian ones?}
}
\value{
A tibble containing the matched substance names and the latest occurrence
in \link{srppp_active_substances_merged}.
}
\description{
Find the year of the last approved use of an active substance
}
\examples{
last_approved("Imidacloprid")
last_approved("Quin")
}
