% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_categories.R
\name{product_categories}
\alias{product_categories}
\title{Get product categories for a pNbr}
\usage{
product_categories(
  pNbr,
  year = format(Sys.Date(), "\%Y"),
  lang = c("de", "fr", "it")
)
}
\arguments{
\item{pNbr}{The P-Number of the product}

\item{year}{The year, defaults to current year}

\item{lang}{The language to be used for the return value}
}
\value{
A character string with the product categories in the SRPPP pasted
together, separated by a comma.
}
\description{
Get product categories for a pNbr
}
\examples{
product_categories(7811)
product_categories(7811, lang = "it")
}
