% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-data.R
\docType{data}
\name{psmv_active_substances}
\alias{psmv_active_substances}
\title{Table of active substance names extracted from all years starting 2011}
\format{
tibble A table, resolving the primary keys ('pk') of the active
substances to the latest entry in any of the 'substances' tables in
\link{psmv_list}.
}
\usage{
psmv_active_substances
}
\description{
Table of active substance names extracted from all years starting 2011
}
\examples{
print(psmv_active_substances, n = Inf)
}
\keyword{datasets}
