% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-data.R
\docType{data}
\name{psmv_list}
\alias{psmv_list}
\title{List of 'psmv_dm' objects for all years starting 2011}
\format{
list A list of PSMV versions created with \link{psmv_dm}, named with a
character vector of the respective years
}
\usage{
psmv_list
}
\description{
For each year, the first XML dump published by FOAG is used.
}
\examples{
names(psmv_list)
}
\keyword{datasets}
