% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-data.R
\docType{data}
\name{psmv_xml_dates}
\alias{psmv_xml_dates}
\title{Publication dates of the available zip files}
\format{
character vector of publication dates in the format YYYY-MM-DD
}
\usage{
psmv_xml_dates
}
\description{
Publication dates of the available zip files
}
\examples{
print(psmv_xml_dates)
}
\keyword{datasets}
