% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-get.R
\name{psmv_xml_get.numeric}
\alias{psmv_xml_get.numeric}
\alias{psmv_xml_get.Date}
\title{Additional methods to read an XML version of the PSMV}
\usage{
\method{psmv_xml_get}{numeric}(from, ...)

\method{psmv_xml_get}{Date}(from, ...)
}
\arguments{
\item{from}{A number giving a year starting from 2011 up to the current year, or
one of the dates in \link{psmv_xml_dates}.}

\item{\dots}{Currently not used}
}
\value{
An object inheriting from 'psmv_xml', 'xml_document', 'xml_node'
}
\description{
Additional methods to read an XML version of the PSMV
}
\examples{
\dontrun{
# The following only works if you have a collection of zipped XML dumps in a directory
# specified in the environment variable R_psmvhist_idir
psmv_2015 <- psmv_xml_get(2015)
print(psmv_2015)
class(psmv_2015)

# This is the method for dates used behind the scenes
psmv_2017 <- psmv_xml_get(as.Date("2017-01-13"))
print(psmv_2017)
}
}
