% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-data.R
\docType{data}
\name{psmv_xml_zip_files}
\alias{psmv_xml_zip_files}
\title{Relative paths of the available zip files}
\format{
character vector of paths relative to 'psmv_xml_idir',
named with their publication dates in the format YYYY-MM-DD
}
\usage{
psmv_xml_zip_files
}
\description{
Relative paths of the available zip files
}
\examples{
print(psmv_xml_zip_files)
}
\keyword{datasets}
