% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmvhist-package.R
\docType{package}
\name{psmvhist-package}
\alias{psmvhist}
\alias{psmvhist-package}
\title{psmvhist: Historical Data from the Swiss Register of Plant Protection Products as an R Package}
\description{
Data objects extracted from historical XML versions of the Swiss Register of Plant Protection Products 'PSMV'. Starting from 2011, one such data object is contained for each year. Note that the information contained in this project is outdated. An online version of the current register can be accessed via \url{https://www.psm.admin.ch/de/produkte}. An XML dump of the current register can be read in using the 'psmv' package.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://gitlab.agsad.admin.ch/psmwb/psmvdata/issues}
}

}
\author{
\strong{Maintainer}: Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID})

Other contributors:
\itemize{
  \item Daniel Baumgartner [contributor]
  \item Agroscope [copyright holder]
}

}
\keyword{internal}
