% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml-get.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{srppp_xml_get}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{srppp}{\code{\link[srppp:srppp_xml_get]{srppp_xml_get()}}}
}}

