% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_compositions}
\alias{srppp_compositions}
\title{Table of the latest product compositions}
\format{
A \link[tibble:tibble]{tibble} with the columns P-Number 'pNbr', substance
primary key 'pk', 'type' and product composition columns 'percent', 'g_per_L',
'ingredient_de', 'ingredient_fr' and 'ingredient_it', and the column
'latest', indicating the year entry of that combination.
}
\usage{
srppp_compositions
}
\description{
This table contains only one composition for each P-Number. In cases, where
several versions of the composition are given in different years, the composition
in the latest year of occurrence is used, assuming that it is the correct one.
}
\examples{
print(srppp_compositions, n = Inf)
}
\seealso{
\link{srppp_ingredients} for all unique compositions of a product

\link{srppp_products}
}
\keyword{datasets}
