% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_ingredients}
\alias{srppp_ingredients}
\title{Table of product ingredients extracted from all years starting 2011}
\format{
A \link[tibble:tibble]{tibble} of all unique combinations of P-Number, substance
primary key 'pk', 'type' and product composition columns 'percent', 'g_per_L',
'ingredient_de', 'ingredient_fr' and 'ingredient_it', and the column
'latest', indicating the latest entry of that combination.
}
\usage{
srppp_ingredients
}
\description{
This table contains all unique ingredient information extracted from
the XML files. Some products have different ingredient information
in different years. In this table, all unique compositions occurring
in one of the years are collected.
}
\examples{
print(srppp_ingredients, n = Inf)
}
\seealso{
\link{srppp_compositions} for one product composition per product

\link{srppp_products}
}
\keyword{datasets}
