% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_substances_merged}
\alias{srppp_substances_merged}
\title{Table of substance names extracted from all years starting 2011}
\format{
A \link[tibble:tibble]{tibble}. The different types of primary keys (integer and uuid)
are listed in the columns \code{pk_v1} and \code{pk_v2}, respectively. The earliest
and the latest year of occurrence are given in the columns 'earliest' and
'latest'.
}
\usage{
srppp_substances_merged
}
\description{
A table of all unique substance names (strictly speaking, of all unique
combinations of German and French names) occurring in any of the 'substances'
tables in \link{srppp_list}.
}
\note{
Note that a substance occurring in the register in a certain year does
not necessarily mean that a product containing the substance was
authorized. In particular, a number of products with such comments are
present in the register used as representative for 2011.
}
\examples{
print(srppp_substances_merged, n = Inf)
}
\seealso{
\link{srppp_substances}, \link{srppp_products}
}
\keyword{datasets}
