% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_xml_dates}
\alias{srppp_xml_dates}
\title{Publication dates of the available zip files}
\format{
character vector of publication dates in the format YYYY-MM-DD
}
\usage{
srppp_xml_dates
}
\description{
Publication dates of the available zip files
}
\examples{
print(srppp_xml_dates)
}
\keyword{datasets}
