% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml-get.R
\name{srppp_xml_get.numeric}
\alias{srppp_xml_get.numeric}
\alias{srppp_xml_get.Date}
\title{Additional methods to read an XML version of the Swiss Register of Plant Protection Products}
\usage{
\method{srppp_xml_get}{numeric}(from, ...)

\method{srppp_xml_get}{Date}(from, ...)
}
\arguments{
\item{from}{A number giving a year starting from 2011 up to the current year, or
one of the dates in \link{srppp_xml_dates}.}

\item{\dots}{Currently not used}
}
\value{
An object inheriting from 'srppp_xml', 'xml_document', 'xml_node'
}
\description{
Additional methods to read an XML version of the Swiss Register of Plant Protection Products
}
\examples{
\dontrun{
# The following only works if you have a collection of zipped XML dumps in a directory
# specified in the environment variable R_srppphist_idir
srppp_2015 <- srppp_xml_get(2015)
print(srppp_2015)
class(srppp_2015)

# This is the method for dates used behind the scenes
srppp_2017 <- srppp_xml_get(as.Date("2017-01-13"))
print(srppp_2017)
}
}
