% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_xml_zip_files}
\alias{srppp_xml_zip_files}
\title{Relative paths of the available zip files}
\format{
character vector of paths relative to 'srppp_xml_idir',
named with their publication dates in the format YYYY-MM-DD
}
\usage{
srppp_xml_zip_files
}
\description{
Relative paths of the available zip files
}
\examples{
print(srppp_xml_zip_files)
}
\keyword{datasets}
