% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-package.R
\docType{package}
\name{srppphist-package}
\alias{srppphist}
\alias{srppphist-package}
\title{srppphist: Historical Data from the Swiss Register of Plant Protection Products as an R Package}
\description{
Data objects extracted from historical XML versions of the Swiss Register of Plant Protection Products 'SRPPP'. Starting from 2011, one such data object is contained for each year. Note that the information contained in this project is outdated. An online version of the current register can be accessed via \url{https://www.psm.admin.ch/de/produkte}. An XML dump of the current register can be read in using the 'srppp' package. There is no guarantee of correspondence of the data contained in or read in using this package with any online version, or with the original registration documents. Also, the Federal Food Safety and Veterinary Office, coordinating the authorisation of plant protection products in Switzerland, does not answer requests regarding this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agroscope-ch.github.io/srppphist}
  \item Report bugs at \url{https://github.com/agroscope-ch/srppphist/issues}
}

}
\author{
\strong{Maintainer}: Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID})

Authors:
\itemize{
  \item Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel Baumgartner [contributor]
  \item Romualdus Kasteel \email{romualdus.kasteel@agroscope.admin.ch} [reviewer]
  \item Marcel Mathis \email{marcel.mathis@agroscope.admin.ch} [contributor, reviewer]
  \item Agroscope [copyright holder]
}

}
\keyword{internal}
