% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{pNbr}
\alias{pNbr}
\title{Retrieve and select a P-Number for a product name}
\usage{
pNbr(string, select = interactive())
}
\arguments{
\item{string}{The search string}
}
\description{
Retrieve and select a P-Number for a product name
}
\examples{
pNbr("Karate")
}
