% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-get.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{psmv_xml_get}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{psmv}{\code{\link[psmv]{psmv_xml_get}}}
}}

