% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_active_substances}
\alias{srppp_active_substances}
\title{Table of active substance names extracted from all years starting 2011}
\format{
tibble A table, resolving the primary keys ('pk') of the active
substances to the latest entry in any of the 'substances' tables in
\link{srppp_list}.
Also, the earliest and the latest year of occurrence
are given in the columns 'earliest' and 'latest'.
}
\usage{
srppp_active_substances
}
\description{
Table of active substance names extracted from all years starting 2011
}
\examples{
print(srppp_active_substances, n = Inf)
}
\keyword{datasets}
