% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_active_substances_merged}
\alias{srppp_active_substances_merged}
\title{Table of active substance names extracted from all years starting 2011}
\format{
A \link[tibble:tibble]{tibble}
}
\usage{
srppp_active_substances_merged
}
\description{
Please note that products whose authorisation has expired remain in the
files until the end of the period during which use by the end user is
still permitted. Therefore, the latest year given in this dataset means
the latest year where a use up period ('expirationDeadline') of at least one
product containing the substance had not expired in the beginning of the
respective year.
}
\details{
Please see \link{srppp_substances} and \link{srppp_substances_merged} for details
on the different primary keys used in these tables.
}
\note{
Note that a substance occurring in the register in a certain year
does not necessarily mean that a product containing the substance was
authorized. In particular, a number of products with such comments are present
in the register used as representative for 2011.
}
\examples{
print(srppp_active_substances_merged, n = Inf)
}
\seealso{
\link{srppp_active_substances}, \link{srppp_substances_merged}, \link{srppp_products}
}
\keyword{datasets}
