% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_ingredients}
\alias{srppp_ingredients}
\title{Table of product ingredients extracted from all years starting 2011}
\format{
tibble A table of all unique combinations of P-Number, substance
primary key 'pk', and product composition columns 'percent', 'g_per_L',
'ingredient_de', 'ingredient_fr' and 'ingredient_it', and the column
'latest', indicating the latest entry of that combination.
}
\usage{
srppp_ingredients
}
\description{
Table of product ingredients extracted from all years starting 2011
}
\examples{
print(srppp_ingredients, n = Inf)
}
\keyword{datasets}
