% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_list}
\alias{srppp_list}
\title{List of 'srppp_dm' objects for all years starting 2011}
\format{
list A named list of \code{srppp::srppp_dm} objects created with the companion
package 'srppp'. The list elements are named with the years from 2011 to the
current year as a character vector
}
\usage{
srppp_list
}
\description{
For each year, the first XML dump published by the registration authority is
used, with few exceptions, where a corrected dump was published shortly
after the first one. Please note the use conditions set out by the
registration authority for the XML dumps currently published at their website.
}
\section{Use conditions set out by the registration authority}{
Please consult the
\href{https://www.blv.admin.ch/blv/de/home/zulassung-pflanzenschutzmittel/pflanzenschutzmittelverzeichnis.html}{use conditions of the XML data files}
currently published by the Federal Food Safety and Veterinary Office (FSVO).
For the the historical data contained in this package, the following points
are of particular importance:
\itemize{
\item In cases of doubt, the definitive source of information are always the
original registration documents, for present as well as past authorisations.
\item Commercial use of the data provided as XML files is not permitted without
the written consent of the FSVO.
}
}

\section{Additional notes regarding proper use of the data}{
As we include only historical, not current authorisation data in this package,
please note the following:
\itemize{
\item The descriptions of products and their authorised uses contained in this
package refer to past authorisations. Regarding current authorisation,
please refer to the \href{https://www.psm.admin.ch/de/produkte}{Swiss Register of Plant Protection Products}, or use the \code{srppp} package
which facilitates reading in the current registration data into R.
\item Products whose authorisation has expired or which have been withdrawn
from the parallel import list are present in the historical data until the
end of the period during which use by the end user is still permitted
('exhaustionDeadline'). This date and the sell-out period
('soldoutDeadline') are indicated in the \code{products} table of each \code{srppp_dm}
object.
\item If you use the historical registration data in the form provided by this
package, please cite the package as described by the output of
\code{citation("srppphist")}.
}
}

\examples{
names(srppp_list)

# In case you are interested in the registered uses of products containing
# a certain active substance, here is some example code
library(dplyr, warn.conflicts = FALSE)
# Step 1: Get the pk number of a certain active substance
pk_active <- srppp_active_substances |>
  filter(substance_de == "Cyproconazole") |>
  pull(pk)

# Step 2: Get the products (pNbrs) containing that substance in 2018
products_2018 <- srppp_list[["2018"]]$ingredients |>
  filter(pk == pk_active)

# Step 3: Get the associated uses
uses_2018 <- products_2018 |>
  left_join(srppp_list[["2018"]]$uses, by = "pNbr")

# Step 4: Add additional information, e.g. the cultures
uses_x_cultures_2018 <- uses_2018 |>
  left_join(srppp_list[["2018"]]$cultures, by = c("pNbr", "use_nr"))

# Step 5: Application rate in g/ha
uses_x_cultures_2018_rate <- uses_x_cultures_2018 |>
  srppp::application_rate_g_per_ha() |>
  select(pNbr, use_nr, application_area_de, culture_de, rate_g_per_ha)

# If this should be repeated for all available years, it is convenient
# to define a function that extracts the desired information, apply it
# to the list of yearly product registers, and combine the results in a
# table.
uses_cultures_rates <- function(sr, pk_active) {
  sr$ingredients |>
    filter(pk == pk_active) |>
    left_join(sr$uses, by = "pNbr") |>
    left_join(sr$cultures, by = c("pNbr", "use_nr")) |>
    srppp::application_rate_g_per_ha() |>
    select(pNbr, use_nr, application_area_de,
      culture_de, rate_g_per_ha)
}

# Test the function
uses_cultures_rates(srppp_list[["2018"]], 116L)

# Create a list of tables
uses_cultures_rates_list <- lapply(srppp_list, uses_cultures_rates, 116L)

# Combine the tables for all years
uses_cultures_rates_all_years <- bind_rows(uses_cultures_rates_list,
  .id = "year")
print(uses_cultures_rates_all_years)

# Find names of original products and sales permissions (W-Numbers with dash)
uses_cultures_rates_all_years |>
  select(year, pNbr) |>
  unique() |>
  left_join(srppp_products[c("pNbr", "wNbr", "name")], by = "pNbr",
    relationship = "many-to-many")
}
\keyword{datasets}
