% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_obligations_spe3}
\alias{srppp_obligations_spe3}
\title{Table of all SPe 3 obligations and corresponding mitigation measures}
\format{
A table with the relevant obligation texts in German, and its
corresponding mitigation measures
}
\usage{
srppp_obligations_spe3
}
\description{
The quantitative mitigation measures extracted from the text
as explained in the documentation to \link[srppp:srppp_dm]{srppp::srppp_dm}.
}
\examples{
print(srppp_obligations_spe3, n = Inf)
}
\keyword{datasets}
