% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_pests}
\alias{srppp_pests}
\title{Table of pesticide targets ("pests") extracted from all years starting 2011}
\format{
tibble A table of all pesticide target names in German, French and
Italian, a scientific name if available and the columns 'earliest' and 'latest',
indicating the earliest and latest year of occurrence.
}
\usage{
srppp_pests
}
\description{
Table of pesticide targets ("pests") extracted from all years starting 2011
}
\examples{
print(srppp_pests, n = Inf)
}
\keyword{datasets}
