% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppphist-data.R
\docType{data}
\name{srppp_products}
\alias{srppp_products}
\title{Table of products extracted from all years starting 2011}
\format{
tibble A table of all unique combinations of P-Number, W-Number
and product name over all years, and the columns 'earliest' and 'latest',
indicating the earliest and latest occurrence of that combination.
}
\usage{
srppp_products
}
\description{
Table of products extracted from all years starting 2011
}
\examples{
print(srppp_products, n = Inf)
}
\keyword{datasets}
